Move VTarget top level into correct folder
